cloud_functions/feature-transform/transform_repos_llm_enrich/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.20.1
google-cloud-storage==2.14.0
pandas==2.2.2
requests==2.31.0
openai==1.23.6
db-dtypes==1.2.0

